requirements.txt
Flask
flask-cors
requests
Pillow\npillow-heif
numpy
opencv-python-headless
ImageHash
pyjwt
bcrypt
SQLAlchemy
python-dotenv
torch
timm

